Trying to revert commit
I was about to deliver, when some of my img disappeared with broken tag, my boxes were all over the place, so I reverted to last commit. Here are all my last changes before the disaster, but still it does not appear right on my project. My imgages, icons, boxes are totally all over the place and I am not sure how to fix this...